requirements.txt
blinker == 1.9.0
certifi == 2025.4.26
charset-normalizer == 3.4.2
click == 8.2.0
colorama == 0.4.6
Flask == 3.1.0
Flask-SQLAlchemy == 3.1.1
greenlet == 3.2.2
idna == 3.10
itsdangerous == 2.2.0
Jinja2 == 3.1.6
MarkupSafe == 3.0.2
nba_api == 1.9.0
numpy == 2.2.5
pandas == 2.2.3
psycopg2-binary == 2.9.10
python-dateutil == 2.9.0.post0
python-dotenv == 1.1.0
pytz == 2025.2
requests == 2.32.3
SQLAlchemy == 2.0.41
typing_extensions == 4.14.0
tzdata == 2025.2
urllib3 == 2.4.0
Werkzeug == 3.1.3
plotly == 6.1.2
